+++ contracts/prod/1.1/base/ERC20interface.sol
@@ -10,25 +10,25 @@ pragma solidity 0.8.20;
  * This contract serves as a foundation for the ERC20Base contract.
  */
 interface ERC20interface {
-    function decimals() external view returns (uint dec);
+    function decimals() external view returns (uint256 dec);
 
-    function totalSupply() external view returns (uint _totalSupply);
+    function totalSupply() external view returns (uint256 _totalSupply);
 
-    function balanceOf(address _owner) external view returns (uint balance);
+    function balanceOf(address _owner) external view returns (uint256 balance);
 
-    function transfer(address _to, uint _value) external returns (bool success);
+    function transfer(address _to, uint256 _value) external returns (bool success);
 
-    function transferFrom(address _from, address _to, uint _value) external returns (bool success);
+    function transferFrom(address _from, address _to, uint256 _value) external returns (bool success);
 
-    function approve(address _spender, uint _value) external returns (bool success);
+    function approve(address _spender, uint256 _value) external returns (bool success);
 
-    function allowance(address _owner, address _spender) external view returns (uint remaining);
+    function allowance(address _owner, address _spender) external view returns (uint256 remaining);
 
     function name() external view returns (string memory);
 
     function symbol() external view returns (string memory);
 
-    event Transfer(address indexed _from, address indexed _to, uint _value);
-    event TransferredFrom(address indexed _from, address indexed _to, uint _value);
-    event Approval(address indexed _owner, address indexed _spender, uint _value);
+    event Transfer(address indexed _from, address indexed _to, uint256 _value);
+    event TransferredFrom(address indexed _from, address indexed _to, uint256 _value);
+    event Approval(address indexed _owner, address indexed _spender, uint256 _value);
 }
