+++ contracts/prod/1.1/base/Interfaces.sol
@@ -4,88 +4,88 @@ pragma solidity 0.8.20;
 import "../../../openzeppelin-v5.0.1/token/ERC20/IERC20.sol";
 
 interface IContractsManager {
-    function votingTokenContracts(uint projectId) external view returns (IERC20);
-    function addBurnAddressProposals(uint projectId, uint proposalId) external view returns (address);
-    function nextProjectId() external view returns (uint);
-    function getBurnAddresses(uint projectId) external view returns (address[] memory);
-    function burnedTokens(uint projectId, IERC20 votingTokenContract) external view returns (uint);
-    function votingTokenCirculatingSupply(uint projectId) external view returns (uint);
+    function votingTokenContracts(uint256 projectId) external view returns (IERC20);
+    function addBurnAddressProposals(uint256 projectId, uint256 proposalId) external view returns (address);
+    function nextProjectId() external view returns (uint256);
+    function getBurnAddresses(uint256 projectId) external view returns (address[] memory);
+    function burnedTokens(uint256 projectId, IERC20 votingTokenContract) external view returns (uint256);
+    function votingTokenCirculatingSupply(uint256 projectId) external view returns (uint256);
     function createProject(string memory dbProjectId, address tokenContractAddress) external;
     function createProject(string memory dbProjectId, address tokenContractAddress, bool checkErc20) external;
-    function proposeAddBurnAddress(uint projectId, address burnAddress) external;
-    function executeProposal(uint projectId, uint proposalId) external;
-    function hasMinBalance(uint projectId, address addr) external view returns (bool);
-    function minimumRequiredAmount(uint projectId) external view returns (uint);
+    function proposeAddBurnAddress(uint256 projectId, address burnAddress) external;
+    function executeProposal(uint256 projectId, uint256 proposalId) external;
+    function hasMinBalance(uint256 projectId, address addr) external view returns (bool);
+    function minimumRequiredAmount(uint256 projectId) external view returns (uint256);
 
-    function burnAddresses(uint projectId, uint index) external view returns (address);
+    function burnAddresses(uint256 projectId, uint256 index) external view returns (address);
 }
 
 interface IDelegates {
-    function getDelegatorsOf(uint projectId, address delegatedAddress) view external returns (address[] memory);
-    function getDelegatedVotingPower(uint projectId, address delegatedAddr) external view returns (uint);
-    function checkVotingPower(uint projectId, address addr, uint minVotingPower) external view returns (bool);
-    function delegate(uint projectId, address delegatedAddr) external;
-    function undelegate(uint projectId) external;
-    function undelegateAllFromAddress(uint projectId) external;
-    function removeSpamDelegates (uint projectId, address[] memory addresses, uint[] memory indexes) external;
-    function getSpamDelegates(uint projectId, address delegatedAddr) external view returns (address  [] memory, uint [] memory);
+    function getDelegatorsOf(uint256 projectId, address delegatedAddress) view external returns (address[] memory);
+    function getDelegatedVotingPower(uint256 projectId, address delegatedAddr) external view returns (uint256);
+    function checkVotingPower(uint256 projectId, address addr, uint256 minVotingPower) external view returns (bool);
+    function delegate(uint256 projectId, address delegatedAddr) external;
+    function undelegate(uint256 projectId) external;
+    function undelegateAllFromAddress(uint256 projectId) external;
+    function removeSpamDelegates (uint256 projectId, address[] memory addresses, uint256[] memory indexes) external;
+    function getSpamDelegates(uint256 projectId, address delegatedAddr) external view returns (address  [] memory, uint256 [] memory);
 
     // Getters for public fields
-    function delegations(uint, address, uint) external view returns (address);
-    function delegateOf(uint, address) external view returns (address);
+    function delegations(uint256, address, uint256) external view returns (address);
+    function delegateOf(uint256, address) external view returns (address);
 }
 
 
 interface IFundsManager {
     function sendOrphanTokensToGitSwarm(address tokenAddress) external;
-    function depositEth(uint projectId) external payable;
-    function depositToken(uint projectId, address tokenAddress, uint amount) external;
-    function transactionProposal(uint projectId, uint proposalId) external view returns (address[] memory token, uint[] memory amount, address[] memory to, uint[] memory depositToProjectId);
-    function proposeTransaction(uint projectId, address[] memory tokenContractAddress, uint[] memory amount, address[] memory to, uint[] memory depositToProjectId) external;
-    function executeProposal(uint projectId, uint proposalId) external;
-    function reclaimFunds(uint projectId, uint votingTokenAmount, address[] memory tokenContractsAddresses) external;
-    function sendToken(uint projectId, address tokenAddress, address receiver, uint amount) external;
-    function sendEther(uint projectId, address payable receiver, uint amount) external;
-    function updateBalance(uint projectId, address tokenAddress, uint amount) external;
-    function balances(uint projectId, address tokenAddress) external view returns (uint);
+    function depositEth(uint256 projectId) external payable;
+    function depositToken(uint256 projectId, address tokenAddress, uint256 amount) external;
+    function transactionProposal(uint256 projectId, uint256 proposalId) external view returns (address[] memory token, uint256[] memory amount, address[] memory to, uint256[] memory depositToProjectId);
+    function proposeTransaction(uint256 projectId, address[] memory tokenContractAddress, uint256[] memory amount, address[] memory to, uint256[] memory depositToProjectId) external;
+    function executeProposal(uint256 projectId, uint256 proposalId) external;
+    function reclaimFunds(uint256 projectId, uint256 votingTokenAmount, address[] memory tokenContractsAddresses) external;
+    function sendToken(uint256 projectId, address tokenAddress, address receiver, uint256 amount) external;
+    function sendEther(uint256 projectId, address payable receiver, uint256 amount) external;
+    function updateBalance(uint256 projectId, address tokenAddress, uint256 amount) external;
+    function balances(uint256 projectId, address tokenAddress) external view returns (uint256);
 }
 
 
 interface IGasStation {
-    function buyGasForProject(uint projectId) payable external;
-    function proposeTransferToGasAddress(uint amount, address to) external;
-    function executeProposal(uint id) external;
+    function buyGasForProject(uint256 projectId) payable external;
+    function proposeTransferToGasAddress(uint256 amount, address to) external;
+    function executeProposal(uint256 id) external;
 }
 
 
 interface IProposal {
-    function proposals(uint, uint) external view returns (uint32, bool, bool, uint64, uint);
-    function nextProposalId(uint) external view returns (uint);
-    function getVotes(uint, uint, address) external view returns (bool, bool);
-    function setActive(uint, uint, bool) external;
-    function setWillExecute(uint, uint, bool) external;
-    function hasVotedAlready(uint, uint, address) external view returns (bool);
-    function vote(uint, uint, bool) external;
-    function deleteProposal(uint, uint) external;
-    function createProposal(uint, uint32, address) external;
-    function checkVoteCount(uint, uint, uint) external view returns (uint, uint, bool);
-    function lockVoteCount(uint, uint) external;
-    function getVoteCount(uint, uint) external view returns (uint, uint);
-    function getDelegatedVotingPowerExcludingVoters(uint, address, uint,  IERC20) external view returns (uint);
-    function calculateTotalVotingPower(uint, address, uint, IERC20) external view returns (uint);
-    function contestProposal(uint projectId, uint proposalId, bool doRecount) external returns (bool);
+    function proposals(uint256, uint256) external view returns (uint32, bool, bool, uint64, uint256);
+    function nextProposalId(uint256) external view returns (uint256);
+    function getVotes(uint256, uint256, address) external view returns (bool, bool);
+    function setActive(uint256, uint256, bool) external;
+    function setWillExecute(uint256, uint256, bool) external;
+    function hasVotedAlready(uint256, uint256, address) external view returns (bool);
+    function vote(uint256, uint256, bool) external;
+    function deleteProposal(uint256, uint256) external;
+    function createProposal(uint256, uint32, address) external;
+    function checkVoteCount(uint256, uint256, uint256) external view returns (uint256, uint256, bool);
+    function lockVoteCount(uint256, uint256) external;
+    function getVoteCount(uint256, uint256) external view returns (uint256, uint256);
+    function getDelegatedVotingPowerExcludingVoters(uint256, address, uint256,  IERC20) external view returns (uint256);
+    function calculateTotalVotingPower(uint256, address, uint256, IERC20) external view returns (uint256);
+    function contestProposal(uint256 projectId, uint256 proposalId, bool doRecount) external returns (bool);
 }
 
 interface IParameters {
-    function changeParameterProposals(uint, uint) external view returns (bytes32, uint);
-    function parameters(uint, bytes32) external view returns (uint);
-    function initializeParameters(uint) external;
-    function neededToContest(uint) external view returns (uint);
-    function proposeParameterChange(uint, bytes32, uint) external;
-    function executeProposal(uint, uint) external;
+    function changeParameterProposals(uint256, uint256) external view returns (bytes32, uint256);
+    function parameters(uint256, bytes32) external view returns (uint256);
+    function initializeParameters(uint256) external;
+    function neededToContest(uint256) external view returns (uint256);
+    function proposeParameterChange(uint256, bytes32, uint256) external;
+    function executeProposal(uint256, uint256) external;
     function gitswarmAddress() external view returns (address);
-    function isTrustedAddress(uint projectId, address trustedAddress) external view returns (bool);
-    function proposeChangeTrustedAddress(uint projectId, address trustedAddress, bool value) external;
-    function trustedAddresses(uint projectId, address trustedAddress) external view returns (bool);
-    function changeTrustedAddressProposals(uint projectId, uint proposalId) external view returns (address trustedAddress, bool value);
+    function isTrustedAddress(uint256 projectId, address trustedAddress) external view returns (bool);
+    function proposeChangeTrustedAddress(uint256 projectId, address trustedAddress, bool value) external;
+    function trustedAddresses(uint256 projectId, address trustedAddress) external view returns (bool);
+    function changeTrustedAddressProposals(uint256 projectId, uint256 proposalId) external view returns (address trustedAddress, bool value);
 }
